LayoutTests/editing/execCommand/set-backColor-with-color-filter-from-scripts.html

<!DOCTYPE html><!-- webkit-test-runner [ enableColorFilter=true ] -->
<html>
<body>
<div id="test" style="-apple-color-filter: apple-invert-lightness()" contenteditable>hello world</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>
Markup.description('Setting the background color should invert the color through -apple-color-filter');

window.getSelection().setPosition(test, 0);
window.getSelection().modify('extend', 'forward', 'word');
window.getSelection().modify('extend', 'forward', 'word');
document.execCommand('backColor', false, '#224433');
Markup.dump('test', `"world" should be #224433 / rgb(${0x22}, ${0x44}, ${0x33})`);

</script>
</body>
</html>
